HACK: drm: Avoid adding CEA modes to the mode list
Something is wrong w/ the db845c when switching
modes that are basically identical. The display will
blank out and not return.

I don't have a good fix for this, so for now, avoid
adding the CEA modes that duplicate the detailed modes
in the EDID and seem to trigger the issue.

Signed-off-by: John Stultz <john.stultz@linaro.org>
Signed-off-by: Amit Pundir <amit.pundir@linaro.org>
Signed-off-by: Sumit Semwal <sumit.semwal@linaro.org>